refactor: flatten FareContractModal for CarnetDetails, FareContractDetails and ReceiptScreen to be on root stack
