--- src/fare-contracts/Details/index.ts
@@ -1,3 +1,0 @@
-export {CarnetDetailsScreen} from './CarnetDetailsScreen';
-export {DetailsScreen} from './DetailsScreen';
-export {ReceiptScreen} from './ReceiptScreen';
