+++ src/fare-contracts/FareContractInfo.tsx
@@ -38,12 +38,12 @@ import {
 import {screenReaderPause} from '@atb/components/text';
 import {useMobileTokenContextState} from '@atb/mobile-token/MobileTokenContext';
 import {useFirestoreConfiguration} from '@atb/configuration/FirestoreConfigurationContext';
-import {FareContractDetail} from '@atb/fare-contracts/Component/FareContractDetail';
-import {WarningMessage} from '@atb/fare-contracts/Component/WarningMessage';
-import {Barcode} from '@atb/fare-contracts/Details/Barcode';
+import {FareContractDetail} from '@atb/fare-contracts/components/FareContractDetail';
+import {WarningMessage} from '@atb/fare-contracts/components/WarningMessage';
+import {Barcode} from '@atb/fare-contracts/details/Barcode';
 import {SectionSeparator} from '@atb/components/sections';
-import {getLastUsedAccess} from './Carnet/CarnetDetails';
-import {InspectionSymbol} from '@atb/fare-contracts/Component/InspectionSymbol';
+import {getLastUsedAccess} from './carnet/CarnetDetails';
+import {InspectionSymbol} from '@atb/fare-contracts/components/InspectionSymbol';
 
 export type FareContractInfoProps = {
   travelRights: PreActivatedTravelRight[];
