+++ src/fare-contracts/carnet/CarnetDetails.tsx
@@ -7,15 +7,15 @@ import {
 import {getValidityStatus} from '@atb/fare-contracts/utils';
 import * as Sections from '@atb/components/sections';
 import {ValidityHeader} from '@atb/fare-contracts/ValidityHeader';
-import {UsedAccessValidityHeader} from '@atb/fare-contracts/Carnet/UsedAccessValidityHeader';
+import {UsedAccessValidityHeader} from '@atb/fare-contracts/carnet/UsedAccessValidityHeader';
 import {ValidityLine} from '@atb/fare-contracts/ValidityLine';
 import {View} from 'react-native';
 import {SectionSeparator} from '@atb/components/sections';
 import {FareContractInfo} from '@atb/fare-contracts/FareContractInfo';
-import {CarnetFooter} from '@atb/fare-contracts/Carnet/CarnetFooter';
+import {CarnetFooter} from '@atb/fare-contracts/carnet/CarnetFooter';
 import React from 'react';
 import {StyleSheet} from '@atb/theme';
-import {UsedAccessStatus} from '@atb/fare-contracts/Carnet/types';
+import {UsedAccessStatus} from '@atb/fare-contracts/carnet/types';
 
 export function CarnetDetails(props: {
   now: number;
