+++ src/fare-contracts/components/FareContractDetail.tsx
