+++ src/fare-contracts/components/FareContractStatusSymbol.tsx
