+++ src/fare-contracts/components/InspectionSymbol.tsx
