+++ src/fare-contracts/components/WarningMessage.tsx
