+++ src/fare-contracts/details/Barcode.tsx
