+++ src/fare-contracts/details/UnknownFareContractDetails.tsx
