+++ src/fare-contracts/index.ts
@@ -1,2 +1,6 @@
 export {FareContractAndReservationsList} from './FareContractAndReservationsList';
 export {SimpleFareContract} from './SimpleFareContract';
+export {CarnetDetailedView} from './carnet/CarnetDetailedView';
+export {DetailsContent} from './details/DetailsContent';
+export {getValidityStatus} from './utils';
+export {getLastUsedAccess} from './carnet/CarnetDetails';
