+++ src/stacks-hierarchy/Root_CarnetDetailsScreen.tsx
@@ -6,17 +6,17 @@ import {useInterval} from '@atb/utils/use-interval';
 import React, {useState} from 'react';
 import {View} from 'react-native';
 import {ScrollView} from 'react-native-gesture-handler';
-import {CarnetDetailedView} from '../Carnet/CarnetDetailedView';
-import {FareContractModalScreenProps} from './types';
+import {CarnetDetailedView} from '@atb/fare-contracts';
+import {RootStackScreenProps} from '@atb/stacks-hierarchy/navigation-types';
 
 export type CarnetDetailsRouteParams = {
   orderId: string;
   isInspectable: boolean;
 };
 
-type Props = FareContractModalScreenProps<'CarnetDetailsScreen'>;
+type Props = RootStackScreenProps<'Root_CarnetDetailsScreen'>;
 
-export function CarnetDetailsScreen({navigation, route}: Props) {
+export function Root_CarnetDetailsScreen({navigation, route}: Props) {
   const styles = useStyles();
   const [now, setNow] = useState<number>(Date.now());
   useInterval(() => setNow(Date.now()), 2500);
@@ -26,7 +26,7 @@ export function CarnetDetailsScreen({navigation, route}: Props) {
 
   const onReceiptNavigate = () =>
     fc &&
-    navigation.push('PurchaseReceipt', {
+    navigation.push('Root_ReceiptScreen', {
       orderId: fc.orderId,
       orderVersion: fc.version,
     });
