+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_RootScreen/Dashboard_RootScreen.tsx
@@ -276,8 +276,8 @@ export const Dashboard_RootScreen: React.FC<RootProps> = ({
               orderId: string,
             ) => {
               if (isCarnet) {
-                return navigation.navigate('Root_FareContractModal', {
-                  screen: 'CarnetDetailsScreen',
+                return navigation.navigate({
+                  name: 'Root_CarnetDetailsScreen',
                   params: {
                     orderId,
                     isInspectable,
@@ -285,8 +285,8 @@ export const Dashboard_RootScreen: React.FC<RootProps> = ({
                 });
               }
 
-              return navigation.navigate('Root_FareContractModal', {
-                screen: 'FareContractDetails',
+              return navigation.navigate({
+                name: 'Root_FareContractDetailsScreen',
                 params: {orderId},
               });
             }}
